+++ gpdb-20160503/core/Models/DescribeSampleDataResponseBody.cs
@@ -40,6 +40,10 @@ namespace AlibabaCloud.SDK.Gpdb20160503.Models
         [Validation(Required=false)]
         public string RequestId { get; set; }
 
+        [NameInMap("SampleDataStatus")]
+        [Validation(Required=false)]
+        public string SampleDataStatus { get; set; }
+
     }
 
 }